Added comments

+++ src/Renderer/ConsoleRenderer.cpp
@@ -46,23 +46,29 @@ CHAR_INFO* Renderer::render(unsigned short resolutionX, unsigned short resolutio
                 }
                 else if (y > FLOOR)
                 {
+                    // Floor rendering
+                    // Calculate vertical angle of the pixel
                     const double VERT_ANGLE = (y - HALF_HEIGHT) / (double)resolutionY * HALF_VER_FOV;
+                    // Calculate ratio of distances between floor texel and wall intersection
                     const double PROJECTION_RATIO = (wallHeight / 2 / tan(VERT_ANGLE)) / PROJECTED_DISTANCE / resolutionX;
+                    // Project the point into world space
                     double floorX = player.getPositionX() + DELTA_X * PROJECTION_RATIO;
                     double floorY = player.getPositionY() + DELTA_Y * PROJECTION_RATIO;
+                    // Find tile
                     int tileIndex = level.floorIndexAt(floorX, floorY);
                     Tile tile = level.floorTileFrom(tileIndex);
+                    // Sample the texture
                     double sampleX = floorY;
                     double sampleY = -floorX;
                     CHAR_INFO texture = tile.sampleTexture(sampleX, sampleY, TileTypes::FLOOR);
-                    // Floor rendering
+                    // Put it
                     characters[y * resolutionX + x] = texture;
                 }
                 else
                 {
                     // Wall rendering
                     // Find the point where the texture should be sampled
-                    double sampleY = ((double)y - CEILING) / (HEIGHT + 1.f);
+                    double sampleY = ((double)y - CEILING) / ((double)HEIGHT + 1);
                     double sampleX = 0;
                     switch (intersection.normalDirection)
                     {
